+++ CHANGELOG.md
@@ -59,6 +59,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added `nix::unistd::unlinkat`.
   ([#753](https://github.com/nix-rust/nix/pull/753))
+- Added `nix::unistd::mkdirat`
+  ([#754](https://github.com/nix-rust/nix/pull/754))
 
 ### Changed
 - Use native `pipe2` on all BSD targets.  Users should notice no difference.